requirements.txt
streamlit == 1.33.0
opencv-python == 4.9.0.80
numpy == 1.26.0
pillow == 10.0.1
paddleocr == 2.8.1
paddlepaddle == 2.6.2
requests == 2.31.0
streamlit-lottie == 0.0.5
